Safari-611.1.10/LayoutTests/editing/style/apply-font-size-to-multiple-nodes.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable>
A man with 2
<span style="font-family: Courier New, Courier">font faces</span><br>
Death comes in all
<span style="font-size: large">font sizes</span>
</div>
<script>

var test = document.getElementById('test');
document.getSelection().selectAllChildren(test);
document.execCommand('fontSize', false, 7);
Markup.description('Test that WebKit does not add multiple span or font elements when applying font-size to multiple nodes.')
Markup.dump(test, 'You should see exactly one font element with size="7" and exactly one span');

</script>
</body>
</html>
